+++ src/features/advanced/index.ts
@@ -29,5 +29,6 @@ import './nots-clean-notifications';
 import './prod-hide-percent';
 import './prodq-hide-government-links';
 import './prodq-shorten-material-links';
+import './shorten-shpt-blck-address';
 import './shpf-hide-sort-options';
 import './wf-workforce-filters/wf-workforce-filters';